+++ web-ui/src/components/replay-components/ReplayControls.tsx
@@ -1,11 +1,11 @@
-import { Box, Fab, Tooltip } from '@mui/material';
+import { Box, Fab, TextField, Tooltip } from '@mui/material';
 import SkipPrevious from '@mui/icons-material/SkipPrevious';
 import SkipNext from '@mui/icons-material/SkipNext';
 import FirstPage from '@mui/icons-material/FirstPage';
 import LastPage from '@mui/icons-material/LastPage';
 import ArrowBack from '@mui/icons-material/ArrowBack';
 import { useTranslation } from 'react-i18next';
-import { useCallback } from 'react';
+import { useCallback, useEffect, useState } from 'react';
 
 interface ReplayControlsProps {
   currentTick: number;
@@ -14,8 +14,41 @@ interface ReplayControlsProps {
   onGoBack: () => void;
 }
 
+const TickInput: React.FC<{
+  maxTick: number | undefined;
+  value: string;
+  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  onBlur: () => void;
+}> = ({ maxTick, ...props }) => (
+  <TextField
+    size="small"
+    type="number"
+    slotProps={{ input: { inputProps: { min: 0, max: maxTick } } }}
+    sx={{
+      width: 45,
+      '& .MuiInputBase-input': {
+        padding: '8px',
+        textAlign: 'center',
+      },
+      '& input[type=number]': {
+        '-moz-appearance': 'textfield',
+      },
+      '& input[type=number]::-webkit-outer-spin-button, & input[type=number]::-webkit-inner-spin-button': {
+        '-webkit-appearance': 'none',
+        margin: 0,
+      },
+    }}
+    {...props}
+  />
+);
+
 export const ReplayControls: React.FC<ReplayControlsProps> = ({ currentTick, maxTick, setCurrentTick, onGoBack }) => {
   const { t } = useTranslation();
+  const [inputValue, setInputValue] = useState(currentTick.toString());
+
+  useEffect(() => {
+    setInputValue(currentTick.toString());
+  }, [currentTick]);
 
   const handleStepBack = useCallback(() => {
     setCurrentTick((prev) => Math.max(0, prev - 1));
@@ -34,6 +67,18 @@ export const ReplayControls: React.FC<ReplayControlsProps> = ({ currentTick, max
       setCurrentTick(maxTick);
     }
   }, [maxTick, setCurrentTick]);
+
+  const handleBlur = useCallback(() => {
+    const num = parseInt(inputValue, 10);
+
+    if (!isNaN(num) && num >= 0) {
+      const validNum = maxTick !== undefined ? Math.min(num, maxTick) : num;
+      setCurrentTick(validNum);
+      setInputValue(validNum.toString());
+    } else {
+      setInputValue(currentTick.toString());
+    }
+  }, [inputValue, maxTick, setCurrentTick, currentTick]);
 
   return (
     <Box
@@ -62,13 +107,26 @@ export const ReplayControls: React.FC<ReplayControlsProps> = ({ currentTick, max
           <SkipPrevious />
         </Fab>
       </Tooltip>
+      <Tooltip title={t('replay.set_tick')}>
+        <TickInput
+          maxTick={maxTick}
+          value={inputValue}
+          onChange={(e) => {
+            const value = e.target.value;
+            if (value.length <= 3) {
+              setInputValue(value);
+            }
+          }}
+          onBlur={handleBlur}
+        />
+      </Tooltip>
       <Tooltip title={t('replay.step_forward')}>
         <Fab size="small" onClick={handleStepForward} disabled={maxTick === undefined || currentTick >= maxTick}>
           <SkipNext />
         </Fab>
       </Tooltip>
       <Tooltip title={t('replay.jump_to_end')}>
-        <Fab size="small" onClick={handleJumpToEnd}>
+        <Fab size="small" onClick={handleJumpToEnd} disabled={maxTick === undefined || currentTick >= maxTick}>
           <LastPage />
         </Fab>
       </Tooltip>